refactor(agent): unify naming on upstream sync (drop the upgrade_sync alias)
