+++ CLAUDE.md
@@ -40,7 +40,7 @@ Client/server architecture. `vestad` daemon runs on the host (manages Docker con
 
 **Migrations** (agent-state, distinct from the events.db schema migrations above): `migrations.py` runs prompt-based migrations — markdown files under `agent/core/migrations/`. On boot each unapplied migration is delivered as a boot turn (see Notification flow); the prompt's final step calls the `mark_migration_applied(name)` tool, which records it in `state.json`. If the agent never marks it (rate limit, crash, hallucinated success) it re-runs next boot, so migration prompts must be idempotent. Fresh agents pre-mark every shipping migration without running it — migrations exist only to converge legacy state on update.
 
-**Upstream sync (agent-branch distribution)**: on each release, CI publishes the complete agent home (`agent/core` incl. its `pyproject.toml`/`uv.lock`, `agent/skills`, `agent/MEMORY.md`) to the append-only `agent-workspace` branch — one commit + one `agent-vX.Y.Z` snapshot tag per release (`tools/publish-agent-branch.sh`, gated on test-live, atomic branch+tag push). A box's `$HOME` is a cone-mode sparse checkout of that branch (`attach.sh` in the `upstream-sync` core skill; the cone = installed skills). After a vestad upgrade the running core version (read from `core/pyproject.toml`) no longer matches the persisted `last_synced_version`, so `upgrade_sync.py` queues an upstream-sync boot turn: the agent rebases its local changes onto `agent-v<version>` and records completion via the `mark_upstream_synced` tool (unmarked turns re-fire every boot). Managed boxes get the engine (= `agent/core/`) as one read-only mount, kept out of the checkout cone; unmanaged boxes (`--no-manage-core-code`) add `agent/core` to their cone and pull core through the same rebase.
+**Upstream sync (agent-branch distribution)**: on each release, CI publishes the complete agent home (`agent/core` incl. its `pyproject.toml`/`uv.lock`, `agent/skills`, `agent/MEMORY.md`) to the append-only `agent-workspace` branch — one commit + one `agent-vX.Y.Z` snapshot tag per release (`tools/publish-agent-branch.sh`, gated on test-live, atomic branch+tag push). A box's `$HOME` is a cone-mode sparse checkout of that branch (`attach.sh` in the `upstream-sync` core skill; the cone = installed skills). After a vestad upgrade the running core version (read from `core/pyproject.toml`) no longer matches the persisted `last_synced_version`, so `upstream_sync.py` queues an upstream-sync boot turn: the agent rebases its local changes onto `agent-v<version>` and records completion via the `mark_upstream_synced` tool (unmarked turns re-fire every boot). Managed boxes get the engine (= `agent/core/`) as one read-only mount, kept out of the checkout cone; unmanaged boxes (`--no-manage-core-code`) add `agent/core` to their cone and pull core through the same rebase.
 
 **Config (two bricks, one home each)**: agent configuration is split by kind so each value has a single owner. **Identity** (`WS_PORT`, `AGENT_TOKEN`, `AGENT_NAME`, `TZ`, tunnel): vestad writes it to `agents/{agent}.env` on host at creation, bind-mounted read-only into the container; the agent never writes it. **Config** (the agent's instance): a single nested store at `~/agent/data/config.json`, modeled by `config.py`'s `VestaConfig`. It holds the active `provider` (a pydantic discriminated union, `ClaudeConfig | OpenRouterConfig`, carrying that provider's model + context + thinking + credential) plus provider-independent prefs (`agent_personality`, `timezone`, `seed_context`) and the `notification_rules` ruleset. The **provider** is its own resource: set/switched via `PUT /provider` (sign in), tweaked via `PATCH /provider` (model/context/thinking), and cleared via `DELETE /provider` (sign out); `PUT /config` writes prefs and `notification_rules` and rejects a `provider` key. Most `PUT /config` fields apply on the next restart, but `notification_rules` apply live (`monitor_loop` re-reads them from the store each tick, so the rules editor and the `notifications` skill need no restart). A one-shot boot migration folds a legacy `notification_policy.json` (old separate file, with per-source defaults translated to trailing rules) into `notification_rules`. The OpenRouter key lives nested in `provider` (redacted `SecretStr` on the wire); the Claude OAuth blob lives in `.credentials.json`, which the `claude` CLI reads and refreshes directly, so it is loaded into the model at boot but never persisted to the store (see the LLM provider section). `VestaConfig` layers pydantic-settings sources, highest first: init args, the config store (so a write overrides anything stale), env (identity + operational scalars only; the provider lives in the store, not env), then the model's own field defaults. A one-shot boot migration relocates any legacy flat keys (`agent_model`/`agent_provider`/`openrouter_key`/...) into the nested `provider`; that is the only place flat↔nested mapping lives. New-agent defaults + the per-provider catalog (model lists, context presets, thinking support) come from `agent/core/manifest.json`, **generated from the provider models** by `agent/generate-manifest.py` (CI checks it's fresh) and served by vestad at `GET /manifest` — the model is the one source, the manifest its projection for the non-Python layers. Built-in prompts (notification_suffix.md, nightly_dream.md, restart.md, etc.) ship in `agent/core/prompts/` and are read from there (`core_prompts_dir`); the agent's MEMORY.md lives at `~/agent/MEMORY.md`.
